SIGSRE-65: Add SLO Bootstrap Guide (#27)
* SIGSRE-65: Add SLO Bootstrap Guide

There is still a link to be filled in for the SLO Reporting content, but
it's a work in progress still. The link now points to a Google Doc
(which is not likely to be available to folks outside Red Hat) where
ideation is being fleshed out.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>

* SIGSRE-65: PR Feedback

Add a note from Jeremy about this doc being a self-contained bootsrappign guide.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>

Co-authored-by: Lisa Seelye <lisa@users.noreply.github.com>