Cleanup pass to consolidate markdownlint and spelling corrections (#37)
* Address markdownlint issues throughout the repository

One spelling error fixed in the bootstrap guide. Another commit will address spelling and other formatting issues in other documents.

* A pass at spelling errors

There is a mix of British and American English (still) and I'm not quite
sure how to handle it; there is no consensus or discussion on the topic
as of yet.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>
Co-authored-by: Lisa Seelye <lisa@users.noreply.github.com>